Add function to correctly pass special ignored elements
This gets the word counts more right than before because it ensures we
are always in the state of having skipped an additional 0+ non-word characters
after the end of the element we wanted to skip.

For example we will always be beyond the first "[[" in a link or the
first "\" in a latex-macro, so the skipping back of 2 or one character
in those checks works.